📝 docs: Add project scaffolding standards guide and update documentation
Added a comprehensive guide on project scaffolding standards for ML/AI projects in the file `docs/how-to-guides/project-scaffolding-standards.md`. This guide outlines the standardized folder and file structure to be followed in all projects, ensuring consistency and readability. It also highlights the importance of shared configuration files in enforcing these standards across the team:

- .vscode folder content
- .github folder content
- .devcontainer folder content
- Sections [tool.poetry.dev-dependencies], [tool.poetry.group.dev.dependencies], [tool.black], [tool.ruff], [tool.pymarkdown], [tool.coverage.report] of the pyproject.toml file.

This structure facilitates better project management, collaboration, and adherence to best practices.

Files changed:
- Added `docs/how-to-guides/project-scaffolding-standards.md`
- Updated `docs/index.md` to include a link to the new guide
- Updated `docs/how-to-guides/toc-how-to-guides.md` to include the new guide in the table of contents
- Updated `mkdocs.yml` to add the new guide to the navigation